# openapi/admin.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ultrafast Gateway Admin API",
    "description": "The admin API of the Ultrafast gateway, served under /api.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "2.0.0-alpha.1"
  },
  "paths": {
...
    "/api/logs": {
      "get": {
        "tags": [
          "logs"
        ],
        "summary": "Newest first. `before` is the id of the last row of the page before.\nFilters only narrow what the caller may see.",
        "description": "What a caller sees depends on who they are. An admin sees every row. A\nteam lead sees the rows of the teams they lead, the rows of the users who\nare members of those teams, and their own. Anyone else sees their own\nrows. A lead's scope uses the CURRENT membership: a row linked to a team\nonly by its user leaves the lead's view when that user leaves the team.\nRows of keys without an owner are visible to admins only.",
        "operationId": "logs_list",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "How many rows to return, 1 to 200. 50 when left out.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "before",
            "in": "query",
            "description": "The id of the last row of the page before.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Only calls at or after this time: RFC 3339, or a date `YYYY-MM-DD` (UTC, from its start).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Only calls at or before this time: RFC 3339, or a date `YYYY-MM-DD` (UTC, to its end).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key_id",
            "in": "query",
            "description": "Only calls made with this key.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "Only calls of this user.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "team_id",
            "in": "query",
            "description": "Only calls of this team.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "model",
            "in": "query",
            "description": "Only calls answered by, or asking for, this model name.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Only calls answered with this HTTP status, 100 to 599.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "errors",
            "in": "query",
            "description": "`true`: only calls answered with a status of 400 or more. Combines with the other filters.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "description": "Only calls with this tag, written `name:value` (the name ends at the first colon). Repeat it to require several tags: all must match.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The calls the caller may see, newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogPage"
                }
              }
            }
          },
          "400": {
            "description": "The request is not of the expected form.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid session or access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Some parameters are not valid; `fields` names each of them.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "token": []
          }
        ]
      }
    },
    "/api/logs/{id}": {
      "get": {
        "tags": [
          "logs"
        ],
        "summary": "One call with its attempts. The same scope as the list applies; a call\nthe caller may not see is answered like one that does not exist. A lead's\nview uses the current team membership, and rows of keys without an owner\nare visible to admins only.",
        "operationId": "logs_view",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the log row.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The call with its attempts.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogDetailView"
                }
              }
            }
          },
          "401": {
            "description": "No valid session or access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "There is no such call, or the caller may not see it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "description": "Every error of `/api` has this shape.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        }
      },
      "ApiErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "A stable name for the error, such as `not_found`."
          },
          "fields": {
            "type": "object",
            "description": "For `validation_failed` only: a message for each field that is not\nvalid.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "message": {
            "type": "string",
            "description": "Text for a person. It may change."
          }
        }
      },
...
      "LogDetailView": {
        "type": "object",
        "description": "A call with the targets it tried, in order. The fields of a\n[`LogView`], and `attempts`.",
        "required": [
          "id",
          "at",
          "key_id",
          "key_name",
          "user_id",
          "user_email",
          "team_id",
          "team_name",
          "requested",
          "endpoint",
          "stream",
          "status",
          "provider",
          "model",
          "input_tokens",
          "output_tokens",
          "cost_micros",
          "priced",
          "cached",
          "estimated",
          "duration_ms",
          "tags",
          "attempts"
        ],
        "properties": {
          "at": {
            "type": "string",
            "description": "UTC, `YYYY-MM-DD HH:MM:SS`."
          },
          "attempts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LogAttempt"
            }
          },
          "cached": {
            "type": "boolean",
            "description": "Answered from the response cache: `cost_micros` is 0 and `priced`\nis true, and the tokens are those of the cached answer, so usage\nreports count them; no provider was called."
          },
          "cost_micros": {
            "type": "integer",
            "format": "int64",
            "description": "Cost in millionths of a dollar; 0 when `priced` is false."
          },
          "duration_ms": {
            "type": "integer",
            "format": "int64"
          },
          "endpoint": {
            "type": "string"
          },
          "estimated": {
            "type": "boolean",
            "description": "The tokens and cost are an estimate: a stream that ended without the\nprovider's report (the caller went away, or an error came after\ncontent was sent) is charged the input of the call and the streamed\ncharacters / 4, and `priced` stays true when the model has a price."
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "input_tokens": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "key_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "key_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "`null` when the key was deleted."
          },
          "model": {
            "type": [
              "string",
              "null"
            ]
          },
          "output_tokens": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "priced": {
            "type": "boolean"
          },
          "provider": {
            "type": [
              "string",
              "null"
            ],
            "description": "The provider that answered; null when nothing answered (the model is then null too)."
          },
          "requested": {
            "type": "string",
            "description": "The model or route name the caller asked for."
          },
          "status": {
            "type": "integer",
            "format": "int64",
            "description": "What the caller was answered."
          },
          "stream": {
            "type": "boolean"
          },
          "tags": {
            "type": "object",
            "description": "The tags of the call: what it sent in `x-uf-tags` overlaid by its\nkey's. Empty when none.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "team_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "team_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "user_email": {
            "type": [
              "string",
              "null"
            ]
          },
          "user_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
      "LogPage": {
        "type": "object",
        "required": [
          "logs"
        ],
        "properties": {
          "logs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LogView"
            }
          }
        }
      },
      "LogView": {
        "type": "object",
        "description": "One logged call, as `/api` shows it.",
        "required": [
          "id",
          "at",
          "key_id",
          "key_name",
          "user_id",
          "user_email",
          "team_id",
          "team_name",
          "requested",
          "endpoint",
          "stream",
          "status",
          "provider",
          "model",
          "input_tokens",
          "output_tokens",
          "cost_micros",
          "priced",
          "cached",
          "estimated",
          "duration_ms",
          "tags"
        ],
        "properties": {
          "at": {
            "type": "string",
            "description": "UTC, `YYYY-MM-DD HH:MM:SS`."
          },
          "cached": {
            "type": "boolean",
            "description": "Answered from the response cache: `cost_micros` is 0 and `priced`\nis true, and the tokens are those of the cached answer, so usage\nreports count them; no provider was called."
          },
          "cost_micros": {
            "type": "integer",
            "format": "int64",
            "description": "Cost in millionths of a dollar; 0 when `priced` is false."
          },
          "duration_ms": {
            "type": "integer",
            "format": "int64"
          },
          "endpoint": {
            "type": "string"
          },
          "estimated": {
            "type": "boolean",
            "description": "The tokens and cost are an estimate: a stream that ended without the\nprovider's report (the caller went away, or an error came after\ncontent was sent) is charged the input of the call and the streamed\ncharacters / 4, and `priced` stays true when the model has a price."
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "input_tokens": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "key_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "key_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "`null` when the key was deleted."
          },
          "model": {
            "type": [
              "string",
              "null"
            ]
          },
          "output_tokens": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "priced": {
            "type": "boolean"
          },
          "provider": {
            "type": [
              "string",
              "null"
            ],
            "description": "The provider that answered; null when nothing answered (the model is then null too)."
          },
          "requested": {
            "type": "string",
            "description": "The model or route name the caller asked for."
          },
          "status": {
            "type": "integer",
            "format": "int64",
            "description": "What the caller was answered."
          },
          "stream": {
            "type": "boolean"
          },
          "tags": {
            "type": "object",
            "description": "The tags of the call: what it sent in `x-uf-tags` overlaid by its\nkey's. Empty when none.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "team_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "team_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "user_email": {
            "type": [
              "string",
              "null"
            ]
          },
          "user_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "uf_session",
        "description": "The session cookie set by signing in."
      },
      "token": {
        "type": "http",
        "scheme": "bearer",
        "description": "An access token, which starts with `uf-at-`."
      }
    }
  },
...
}